Migrate stripe helpers
