+++ apps/web/pages/api/cancel.ts
@@ -19,8 +19,8 @@ import { isPrismaObjOrUndefined, parseRecurringEvent } from "@calcom/lib";
 import { HttpError } from "@calcom/lib/http-error";
 import { defaultHandler, defaultResponder } from "@calcom/lib/server";
 import prisma, { bookingMinimalSelect } from "@calcom/prisma";
+import { refund } from "@calcom/stripe/server";
 import type { CalendarEvent } from "@calcom/types/Calendar";
-import { refund } from "@ee/lib/stripe/server";
 import { deleteScheduledEmailReminder } from "@ee/lib/workflows/reminders/emailReminderManager";
 import { sendCancelledReminders } from "@ee/lib/workflows/reminders/reminderScheduler";
 import { deleteScheduledSMSReminder } from "@ee/lib/workflows/reminders/smsReminderManager";
